requirements.txt
python-telegram-bot==20.7
google-generativeai>=0.3.0
python-dotenv>=1.0.0
reportlab>=4.0.0
Flask>=2.3.0
requests>=2.31.0
numpy>=1.24.0
Pillow>=10.0.0
